requirements.txt
azure-functions
praw==6.5.1
tweepy==3.8.0
ffmpeg-python==0.2.0